requirements.txt
PyGithub==1.55
vkbottle==3.0.2
